Add files via upload
A botnet is a huge group of devices that have been hacked through the Internet, each of which is called a bot that serves a botnet component or the so-called bot master. A bot uses a command and control channel to manage its network and carry out its attacks. This botnet name derives from the word “botnets” as devices serve the bot without choosing it, just like bots